Changed the accountpage to have a separate component for password changing
